Remove Vault token management and add simple config server refresher
