+++ src/main/resources/application.yml
@@ -19,12 +19,13 @@ spring:
             bucket: ${S3_BUCKET}
             order: 2
         vault:
-          # Keep using AppRole in config, but our refresher will inject fresh tokens
           authentication: APPROLE
           app-role:
             role-id: ${VAULT_ROLE_ID}
             secret-id: ${VAULT_SECRET_ID}
           uri: ${VAULT_URI:http://localhost:8200}
+          # Add properties to handle batch token expiration
+          fail-fast: false
 
 # Management endpoints
 management:
